pyproject.toml
[project]
name = "agentlab"
version = "1.0.0"
description = "Product-grade agent optimization platform — CLI, API, and web console for iterating ADK agent quality"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "jinja2>=3.1.0",
    "uvicorn[standard]>=0.29.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
]

[project.optional-dependencies]
adk = [
    "google-adk>=0.3.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "playwright>=1.40.0",
]

[project.scripts]
agentlab = "runner:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["runner"]

[tool.setuptools.packages.find]
include = [
    "a2a",
    "a2a.*",
    "adk",
    "adk.*",
    "agent",
    "agent.*",
    "agent_skills",
    "agent_skills.*",
    "api",
    "api.*",
    "assistant",
    "assistant.*",
    "builder",
    "builder.*",
    "cicd",
    "cicd.*",
    "cli",
    "cli.*",
    "collaboration",
    "collaboration.*",
    "context",
    "context.*",
    "control",
    "control.*",
    "core",
    "core.*",
    "cx_studio",
    "cx_studio.*",
    "data",
    "data.*",
    "deployer",
    "deployer.*",
    "evals",
    "evals.*",
    "graders",
    "graders.*",
    "judges",
    "judges.*",
    "logger",
    "logger.*",
    "mcp_server",
    "mcp_server.*",
    "multi_agent",
    "multi_agent.*",
    "notifications",
    "notifications.*",
    "observer",
    "observer.*",
    "optimizer",
    "optimizer.*",
    "policy_opt",
    "policy_opt.*",
    "registry",
    "registry.*",
    "rewards",
    "rewards.*",
    "shared",
    "shared.*",
    "simulator",
    "simulator.*",
    "stores",
    "stores.*",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
